+++ src/update-function/update-array.ts
@@ -1,5 +1,5 @@
-import { computeArrayState, FormArrayState, InferredControlState } from '../state';
-import { ProjectFn2 } from './util';
+import { computeArrayState, FormArrayState, InferredControlState, isFormState } from '../state';
+import { ensureState, ProjectFn2 } from './util';
 
 function updateArrayControlsState<TValue>(updateFn: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>) {
   return (state: FormArrayState<TValue>) => {
@@ -39,7 +39,28 @@ function updateArraySingle<TValue>(updateFn: ProjectFn2<InferredControlState<TVa
  * ```
  */
 export function updateArray<TValue>(
+  updateFn: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>,
   ...updateFnArr: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[]
+): (state: FormArrayState<TValue>) => FormArrayState<TValue>;
+
+/**
+ * This update function takes an array of update functions and returns
+ * a projection function that applies all update functions one after another to
+ * a form array state.
+ *
+ * The following (contrived) example uses this function to validate all its
+ * children to be required and mark them as dirty.
+ *
+ * ```typescript
+ * const arrayUpdateFn = updateArray<string>(
+ *   validate(required),
+ *   markAsDirty,
+ * );
+ * const updatedState = arrayUpdateFn(state);
+ * ```
+ */
+export function updateArray<TValue>(
+  updateFnArr: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[],
 ): (state: FormArrayState<TValue>) => FormArrayState<TValue>;
 
 /**
@@ -59,19 +80,44 @@ export function updateArray<TValue>(
  */
 export function updateArray<TValue>(
   state: FormArrayState<TValue>,
+  updateFn: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>,
   ...updateFnArr: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[]
 ): FormArrayState<TValue>;
 
+/**
+ * This update function takes a form array state and an array of update
+ * functions applies all update functions one after another to the state.
+ *
+ * The following (contrived) example uses this function to validate all its
+ * children to be required and mark them as dirty.
+ *
+ * ```typescript
+ * const updatedState = updateArray<string>(
+ *   state,
+ *   validate(required),
+ *   markAsDirty,
+ * );
+ * ```
+ */
 export function updateArray<TValue>(
-  stateOrFunction: FormArrayState<TValue> | ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>,
-  ...updateFnArr: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[]
+  state: FormArrayState<TValue>,
+  updateFnArr: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[],
+): FormArrayState<TValue>;
+
+export function updateArray<TValue>(
+  stateOrFunctionOrFunctionArray:
+    | FormArrayState<TValue>
+    | ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>
+    | ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[],
+  updateFnOrUpdateFnArr?: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>> | ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[],
+  ...rest: ProjectFn2<InferredControlState<TValue>, FormArrayState<TValue>>[]
 ) {
-  if (typeof stateOrFunction !== 'function') {
-    const [first, ...rest] = updateFnArr;
-    return updateArray(first, ...rest)(stateOrFunction);
+  if (isFormState(stateOrFunctionOrFunctionArray)) {
+    const updateFnArr = Array.isArray(updateFnOrUpdateFnArr) ? updateFnOrUpdateFnArr : [updateFnOrUpdateFnArr!];
+    return updateFnArr.reduce((s, updateFn) => updateArraySingle<TValue>(updateFn)(s), stateOrFunctionOrFunctionArray);
   }
 
-  return (state: FormArrayState<TValue>): FormArrayState<TValue> => {
-    return [stateOrFunction as any, ...updateFnArr].reduce((s, updateFn) => updateArraySingle<TValue>(updateFn)(s), state);
-  };
+  let updateFnArr = Array.isArray(stateOrFunctionOrFunctionArray) ? stateOrFunctionOrFunctionArray : [stateOrFunctionOrFunctionArray];
+  updateFnArr = updateFnOrUpdateFnArr === undefined ? updateFnArr : updateFnArr.concat(updateFnOrUpdateFnArr);
+  return (s: FormArrayState<TValue>) => updateArray<TValue>(ensureState(s), updateFnArr.concat(rest));
 }